+++ Tests/Test_SetS_Tides/Test_Unique_Modes/test_unique_freqs.py
@@ -48,11 +48,13 @@ def test_find_unique_frequency_with_zeros():
     assert len(unique_frequency_list) == len(example_dict) - 2
     assert len(unique_frequency_dict) == len(example_dict) - 2
     assert len(unique_frequency_count_list) == len(example_dict) - 2
-    for i, (sig, freq) in enumerate(example_dict.items()):
-        if freq == 0.:
-            assert 0. not in unique_frequency_list
-            assert sig not in unique_frequency_dict
-            continue
+
+    assert 0. not in unique_frequency_list
+    assert 0. not in unique_frequency_list2
+    assert 'o' not in unique_frequency_dict
+    assert '2n-2o' not in unique_frequency_dict
+
+    for i, (sig, freq) in enumerate([('n', 0.3), ('n-o', 0.5)]):
         assert freq in unique_frequency_list
         assert sig in unique_frequency_dict
         assert unique_frequency_dict[sig] == freq
@@ -76,18 +78,19 @@ def test_find_unique_frequency_with_exact_dups():
     assert len(unique_frequency_list) == len(example_dict) - 1
     assert len(unique_frequency_dict) == len(example_dict) - 1
     assert len(unique_frequency_count_list) == len(example_dict) - 1
-    for i, (sig, freq) in enumerate(example_dict.items()):
-        if freq == 0.:
-            continue
+
+    assert 'o' not in unique_frequency_dict
+    assert '2n-2o' not in unique_frequency_dict
+    assert 'o 2n-2o' in unique_frequency_dict
+
+    for i, (sig, freq) in enumerate([('n', 0.3), ('o 2n-2o', 0.4), ('n-o', 0.5)]):
         assert freq in unique_frequency_list
-        if sig in ['o', '2n-2o']:
-            assert 'o 2n-2o' in unique_frequency_dict
-            assert unique_frequency_dict['o 2n-2o'] == freq
+        assert sig in unique_frequency_dict
+        assert unique_frequency_dict[sig] == freq
+        if i == 1:
             assert unique_frequency_count_list[i] == 2
             assert unique_frequency_count_list2[i] == 2
         else:
-            assert sig in unique_frequency_dict
-            assert unique_frequency_dict[sig] == freq
             assert unique_frequency_count_list[i] == 1
             assert unique_frequency_count_list2[i] == 1
 
@@ -98,7 +101,8 @@ def test_find_unique_frequency_with_near_dups():
     example_dict['n'] = 0.3
     example_dict['o'] = 0.4
     example_dict['n-o'] = 0.5
-    example_dict['2n-2o'] = 0.4 + 10 * MIN_FREQ
+    large_diff = 0.4 + 10 * MIN_FREQ
+    example_dict['2n-2o'] = large_diff
 
     unique_frequency_list, unique_frequency_count_list = find_unique_frequency_list(example_dict)
     unique_frequency_dict, unique_frequency_list2, unique_frequency_count_list2 = find_unique_frequency_dict(example_dict)
@@ -108,11 +112,8 @@ def test_find_unique_frequency_with_near_dups():
     assert len(unique_frequency_list) == len(example_dict)
     assert len(unique_frequency_dict) == len(example_dict)
     assert len(unique_frequency_count_list) == len(example_dict)
-    for i, (sig, freq) in enumerate(example_dict.items()):
-        if freq == 0.:
-            assert 0. not in unique_frequency_list
-            assert sig not in unique_frequency_dict
-            continue
+
+    for i, (sig, freq) in enumerate([('n', 0.3), ('o', 0.4), ('n-o', 0.5), ('2n-2o', large_diff)]):
         assert freq in unique_frequency_list
         assert sig in unique_frequency_dict
         assert unique_frequency_dict[sig] == freq
@@ -126,7 +127,8 @@ def test_find_unique_frequency_with_close_not_exact_dups():
     example_dict['n'] = 0.3
     example_dict['o'] = 0.4
     example_dict['n-o'] = 0.5
-    example_dict['2n-2o'] = 0.4 + 1 * MIN_FREQ
+    small_diff = 0.4 + 1 * MIN_FREQ
+    example_dict['2n-2o'] = small_diff
 
     unique_frequency_list, unique_frequency_count_list = find_unique_frequency_list(example_dict)
     unique_frequency_dict, unique_frequency_list2, unique_frequency_count_list2 = find_unique_frequency_dict(example_dict)
@@ -136,22 +138,18 @@ def test_find_unique_frequency_with_close_not_exact_dups():
     assert len(unique_frequency_list) == len(example_dict) - 1
     assert len(unique_frequency_dict) == len(example_dict) - 1
     assert len(unique_frequency_count_list) == len(example_dict) - 1
-    for i, (sig, freq) in enumerate(example_dict.items()):
-        if freq == 0.:
-            assert 0. not in unique_frequency_list
-            assert 0. not in unique_frequency_list2
-            assert sig not in unique_frequency_dict
-            continue
-        if sig == '2n-2o':
-            assert freq not in unique_frequency_list
-            assert freq not in unique_frequency_list2
-        elif sig == 'o':
-            assert freq in unique_frequency_list
-            assert freq in unique_frequency_list2
+
+    assert 'o' not in unique_frequency_dict
+    assert '2n-2o' not in unique_frequency_dict
+    assert 'o 2n-2o' in unique_frequency_dict
+
+    for i, (sig, freq) in enumerate([('n', 0.3), ('o 2n-2o', 0.4), ('n-o', 0.5)]):
+        assert freq in unique_frequency_list
+        assert sig in unique_frequency_dict
+        assert unique_frequency_dict[sig] == freq
+        if i == 1:
             assert unique_frequency_count_list[i] == 2
             assert unique_frequency_count_list2[i] == 2
         else:
-            assert freq in unique_frequency_list
-            assert freq in unique_frequency_list2
-    assert 'o 2n-2o' in unique_frequency_dict
-    assert unique_frequency_dict['o 2n-2o'] == example_dict['o']
+            assert unique_frequency_count_list[i] == 1
+            assert unique_frequency_count_list2[i] == 1