+++ TidalPy/tides/modes/multilayer_modes.py
@@ -19,6 +19,8 @@ from ..potential import (tidal_potential_nsr, tidal_potential_nsr_modes,
                          tidal_potential_gen_obliquity_nsr_modes, tidal_potential_gen_obliquity_nsr,
                          tidal_potential_obliquity_nsr, tidal_potential_obliquity_nsr_modes, tidal_potential_simple)
 
+from ..heating import calculate_volumetric_heating
+
 if TYPE_CHECKING:
     from ..potential import TidalPotentialOutput
 
@@ -558,27 +560,7 @@ def collapse_multilayer_modes(
     # This stress/strain term is used in the final calculation of tidal heating. It tracks a different coefficient
     #   for frequency averaging (if stress and strain are used on their own then heating would be prop to T^{-2}
     #   rather than T^{-1}
-    # Heating is equal to imag[o] * real[s] - real[o] * imag[s] but we need to multiply by two for the cross terms
-    #    since it is part of a symmetric matrix but only one side of the matrix is calculated in the previous steps.
-    volumetric_heating = (
-        # Im[s_rr] Re[e_rr] - Re[s_rr] Im[e_rr]
-        np.imag(stresses[0]) * np.real(strains[0]) - np.real(stresses[0]) * np.imag(strains[0]) +
-        # Im[s_thth] Re[e_thth] - Re[s_thth] Im[e_thth]
-        np.imag(stresses[1]) * np.real(strains[1]) - np.real(stresses[1]) * np.imag(strains[1]) +
-        # Im[s_phiphi] Re[e_phiphi] - Re[s_phiphi] Im[e_phiphi]
-        np.imag(stresses[2]) * np.real(strains[2]) - np.real(stresses[2]) * np.imag(strains[2]) +
-        # Im[s_rth] Re[e_rth] - Re[s_rth] Im[e_rth]
-        2. * (np.imag(stresses[3]) * np.real(strains[3]) - np.real(stresses[3]) * np.imag(strains[3])) +
-        # Im[s_rphi] Re[e_rphi] - Re[s_rphi] Im[e_rphi]
-        2. * (np.imag(stresses[4]) * np.real(strains[4]) - np.real(stresses[4]) * np.imag(strains[4])) +
-        # Im[s_thphi] Re[e_thphi] - Re[s_thphi] Im[e_thphi]
-        2. * (np.imag(stresses[5]) * np.real(strains[5]) - np.real(stresses[5]) * np.imag(strains[5]))
-    )
-
-
-    # TODO: Without this abs term the resulting heating maps are very blotchy around
-    #    Europa book does have an abs at Equation 42, Page 102
-    volumetric_heating = np.abs(volumetric_heating)
+    volumetric_heating = calculate_volumetric_heating(stresses, strains)
 
     # Perform orbital averaging
     if orbit_average_results: